+++ src/LibHac/Npdm/Acid.cs
@@ -3,6 +3,7 @@ using System;
 using System.IO;
 using LibHac.Common;
 using LibHac.Common.Keys;
+using LibHac.Tools.Crypto;
 
 namespace LibHac.Npdm;
