+++ src/LibHac/Tools/FsSystem/NcaUtils/NcaHeader.cs
@@ -8,6 +8,7 @@ using LibHac.Crypto;
 using LibHac.Diag;
 using LibHac.Fs;
 using LibHac.FsSystem;
+using LibHac.Tools.Crypto;
 using LibHac.Util;
 
 namespace LibHac.Tools.FsSystem.NcaUtils;
